crud de category e entry funcionando

+++ src/main/java/com/trilhajava/trilhajava/services/EntryServiceImpl.java
@@ -2,13 +2,17 @@ package com.trilhajava.trilhajava.services;
 
 import com.trilhajava.trilhajava.dto.EntryDTO;
 import com.trilhajava.trilhajava.entity.EntryEntity;
+import com.trilhajava.trilhajava.exceptions.EntryNotFoundException;
 import com.trilhajava.trilhajava.repositories.EntryRepository;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+import static com.trilhajava.trilhajava.dto.EntryDTO.mapToEntity;
+import static com.trilhajava.trilhajava.entity.EntryEntity.mapToDTO;
 
 @Service
 public class EntryServiceImpl implements EntryService {
@@ -17,8 +21,8 @@ public class EntryServiceImpl implements EntryService {
     private EntryRepository entryRepository;
 
 
-    @Autowired
-    private ModelMapper modelMapper;
+    //@Autowired
+    //private ModelMapper modelMapper;
 
 
     @Override
@@ -39,13 +43,20 @@ public class EntryServiceImpl implements EntryService {
 
     @Override
     public void delete(Long id) {
-        entryRepository.deleteById(id);
+        Optional<EntryEntity> opt = entryRepository.findById(id);
+        try {
+            if (opt.isPresent()) {
+                entryRepository.deleteById(id);
+            } else {
+                throw new EntryNotFoundException("id não encontrado");
+            }
+        } catch (EntryNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        //entryRepository.deleteById(id);
     }
 
-    @Override
-    public void updateById(Long id, EntryDTO dto) {
-
-    }
 
     @Override
     public EntryEntity put(EntryDTO dto) {
@@ -53,7 +64,16 @@ public class EntryServiceImpl implements EntryService {
         //return entryRepository.save(dto);
     }
 
+    @Override
+    public EntryEntity updateById(EntryDTO dto) {
+        EntryEntity entryUpdated = entryRepository.findById(dto.getId())
+                .orElseThrow(() -> new EntryNotFoundException("Entry não encontrada"));
+        entryUpdated.setName(dto.getName());
+        entryUpdated.setDescription(dto.getDescription());
+        return entryRepository.save(entryUpdated);
+    }
 
+    /*
     private EntryEntity mapToEntity(EntryDTO entryDTO) {
         return modelMapper.map(entryDTO, EntryEntity.class);
     }
@@ -61,5 +81,6 @@ public class EntryServiceImpl implements EntryService {
     private EntryDTO mapToDTO(EntryEntity entryEntity) {
         return modelMapper.map(entryEntity, EntryDTO.class);
     }
+     */
 
 }
